correct implementation FacilityServiceImpl class

+++ src/main/java/com/github/fredO1211/booking/repository/FacilityRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FacilityRepository extends JpaRepository<Facility, Integer> {
+public interface FacilityRepository extends JpaRepository<Facility, Long> {
     @Query(nativeQuery = true, value = "SELECT count(*)>0 FROM FACILITIES WHERE NAME=:name")
     boolean isFacilityExists(String name);
 }
